hdayplanner/requirements.txt
Flask
python-dotenv
qrcode
requests
pillow
